pyproject.toml
[tool.poetry]
name = "bib2yaml"
version = "0.1.0"
description = ""
authors = ["Gejza Jenča <gejza.jenca@stuba.sk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pybtex = "^0.24.0"
pyyaml = "^6.0.2"
setuptools = "^75.3.0"

[tool.poetry.scripts]
bib2yaml = "bib2yaml.bib2yaml:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
